refactor: move AuthLoadingScreen to UI library directory
- Relocated AuthLoadingScreen component to src/components/ui/
- Updated import paths in page.tsx and dashboard/layout.tsx
- Verified application builds and runs correctly
- Completes task T007
